feat: 활동 상태관리 서비스 분리 완료
event-emitter를 사용해 의존성 없는 방식으로 구현
