+++ src/app.module.ts
@@ -7,6 +7,7 @@ import { ActivityModule } from './activity/activity.module';
 import { AuthModule } from './auth/auth.module';
 import { ConfigModule } from '@nestjs/config';
 import { SupabaseModule } from './database/supabase.module';
+import { EventEmitterModule } from '@nestjs/event-emitter';
 
 @Module({
   imports: [
@@ -14,6 +15,7 @@ import { SupabaseModule } from './database/supabase.module';
       envFilePath: '.env.development',
       isGlobal: true // 앱 전체에서 ConfigService를 사용할 수 있도록 설정
     }),
+    EventEmitterModule.forRoot(), // 이벤트 발행/구독 기능 활성화
     ClassroomModule, ChatModule, ActivityModule, AuthModule, SupabaseModule
   ],
   controllers: [AppController],
