+++ src/classroom/classroom.service.ts
@@ -3,25 +3,25 @@ import { Classroom } from './classroom.interface';
 import { WsException } from '@nestjs/websockets';
 import { Participant } from './Participant.interface';
 import { Server } from 'socket.io';
-import { Activity } from 'src/activity/activity.interface';
 import { SupabaseService } from 'src/database/supabase.service';
 import { SupabaseClient } from '@supabase/supabase-js';
 import { events } from 'src/utils/events';
+import { EventEmitter2 } from '@nestjs/event-emitter';
 
 const MANAGER_RECONNECT_TIMEOUT = 60000; // 1분
 
 @Injectable()
 export class ClassroomService {
-    private roomData = new Map<string, Classroom>(); // key: classroomId
+    private roomData = new Map<string, Classroom>(); // key: classroomId // 강의실 데이터
     private roomCodeMap = new Map<string, string>(); // key: roomCode, value: classroomId
     private userRoomMap = new Map<string, string>(); // key: socketId, value: classroomId
     private roomRecoveryTimers = new Map<string, NodeJS.Timeout>(); // key: classroomId, value: timerId
 
-    private activities = new Map<string, Activity>(); //key: classroomId 문제풀이활동 세션을 관리하기 위함
-
     private readonly supabase: SupabaseClient; // supabase 클라이언트를 담을 변수
 
-    constructor(private readonly supabaseService: SupabaseService) {
+    constructor(
+        private readonly eventEmitter: EventEmitter2, // 이벤트 발행을 위한 EventEmitter2
+        private readonly supabaseService: SupabaseService) {
         this.supabase = this.supabaseService.getClient(); // Supabase 클라이언트 초기화
     }
 
@@ -47,8 +47,8 @@ export class ClassroomService {
             socketId: managerSocketId,
         });
 
-        // 방 생성 시 해당 방의 활동 상태도 같이 초기화
-        this.initializeActivity(id);
+        // 방 생성 시 해당 방의 활동 상태도 같이 초기화 - eventEmitter를 이용해 의존없는 방식으로 구현
+        this.eventEmitter.emit('room.created', { roomId: id });
 
         this.roomData.set(id, newRoom);
         this.roomCodeMap.set(code, id); // 방 코드와 ID 매핑
@@ -193,10 +193,12 @@ export class ClassroomService {
         }
     }
 
+    // 유예 기간이 활성화되어 있는지 확인
     isGracePeriodActive(classroomId: string): boolean {
         return this.roomRecoveryTimers.has(classroomId);
     }
 
+    // 유예 기간 시작
     startManagerGracePeriod(classroomId: string, managerId: string, server: Server) {
         console.log(`[ClassroomService] Starting ${MANAGER_RECONNECT_TIMEOUT / 1000}s grace period for manager ${managerId} in room ${classroomId}.`);
         const timerId = setTimeout(async () => {
@@ -263,84 +265,6 @@ export class ClassroomService {
     }
 
     // --- 활동 관련 메소드 ---
-
-    // 특정 강의실에 대한 문제풀이 활동 세션을 초기화 함
-    initializeActivity(classroomId: string) {
-        if (!this.activities.has(classroomId)) {
-            this.activities.set(classroomId, {
-                classroomId: classroomId,
-                status: 'waiting',
-                currentQuest: null,
-                partAssignments: [],
-                submissions: {},
-            });
-            console.log(`[ClassroomService: Activity] Initialized activity for classroom ${classroomId}`);
-        }
-    }
-
-    // 선택된 문제를 활동 상태에 저장함
-    setSelectedQuest(classroomId: string, questDetails: any) {
-        const activity = this.activities.get(classroomId);
-        if (activity) {
-            activity.currentQuest = questDetails;
-            activity.status = 'waiting'; // 문제가 선택되었으므로 '시작 대기' 상태
-            console.log(`[ClassroomService] Quest ${questDetails.quest_id} selected for room ${classroomId}`);
-        } else {
-            console.error(`[ClassroomService] Failed to set quest. Activity state not found for room ${classroomId}`);
-        }
-    }
-
-    // 활동을 'active' 상태로 변경하고, 파트 배정 정보를 저장함
-    startActivity(classroomId: string, assignments: any[]): Activity | null{
-        const activity = this.activities.get(classroomId);
-        if (activity && activity.status === 'waiting') {
-            activity.status = 'active'; // 'waiting' -> 'active'
-            activity.partAssignments = assignments;
-            console.log(`[ClassroomService] Activity status changed to 'active' for room ${classroomId}.`);
-            return activity;
-        }
-        return null; // 활동을 시작할 수 없는 경우(이미 시작 되었거나 종료됨)
-    }
-
-    // 특정 사용자의 제출물을 활동상태에 저장/업데이트 함
-    updateUserSubmission(classroomId: string, userId: string, partNumber: number, submissionContent: any): Activity | null {
-        const activity = this.activities.get(classroomId);
-        if (activity && activity.status === 'active') { // 활동이 'active' 상태일 때만 제출 가능
-            // 제출물 업데이트
-            if (!activity.submissions[userId]) {
-                activity.submissions = {};
-            }
-            activity.submissions[userId] = {
-                partNumber: partNumber,
-                content: submissionContent,
-            };
-            console.log(`[ClassroomService] Submission updated for user ${userId} in room ${classroomId}.`);
-            return activity;
-        }
-        return null; // 제출할 수 없는 경우
-    }
-
-    // 해당 강의실의 활동 상태를 초기화 함(문제 선택 전 상태로 되돌림)
-    endCurrentActivity(classroomId: string): boolean {
-        const activity = this.activities.get(classroomId);
-        // 활동이 'active' 또는 'waiting' 일 때만 초기화 의미가 있음
-        if (activity && (activity.status === 'active' || activity.status === 'waiting')) {
-            activity.status = 'waiting'; // 상태를 'waiting'으로 되돌림
-            activity.currentQuest = null; // 현재 문제 세트 초기화
-            activity.partAssignments = []; // 파트 배정 초기화
-            activity.submissions = {}; // 제출물 초기화
-            console.log(`[ClassroomService] Activity ended and reset for room ${classroomId}.`);
-            return true;
-        }
-        // 종료할 활동이 없는 경우 (이미 초기 상태)
-        return false;
-    }
-
-    // 활동 상태를 가져오는 헬퍼 메소드
-    getActivityState(classroomId: string): Activity | undefined {
-        return this.activities.get(classroomId);
-    }
-
     // socketId로 방 ID를 찾는 헬퍼 메소드
     getRoomIdBySocketId(socketId: string): string | undefined {
         return this.userRoomMap.get(socketId);
@@ -351,12 +275,5 @@ export class ClassroomService {
         return this.roomData.get(classroomId); // 방 ID로 방 정보 찾기
     }
 
-    // 해당하는 방의 모든 제출물을 가져옴
-    getAllSubmissions(classroomId: string): Record<string, any> {
-        const activity = this.activities.get(classroomId);
-        if (activity) {
-            return activity.submissions;
-        }
-        throw new WsException('활동 정보가 없습니다.');
-    }
+    
 }
